+++ Frameworks/ZiggeoMediaSDK.framework/Headers/ZiggeoAudios.h
@@ -1,6 +1,6 @@
 //
 //  ZiggeoAudios.h
-//  ZiggeoFramework
+//  ZiggeoMediaSDK
 //
 //  Created by dragon on 11/15/18.
 //  Copyright © 2016 Ziggeo Inc. All rights reserved.