+++ Examples/ZiggeoDemo/ZiggeoDemo/ViewControllers/Home/HomeViewController.swift
@@ -141,7 +141,7 @@ class HomeViewController: UIViewController {
             }
                        
             if (tokens.count > 0) {
-                Common.ziggeo?.playVideos(tokens)
+                Common.ziggeo?.playVideo(tokens)
             } else {
                 Common.showAlertView("Video recordings are empty.")
             }
@@ -155,7 +155,7 @@ class HomeViewController: UIViewController {
             }
             
             if (tokens.count > 0) {
-                Common.ziggeo?.playAudios(tokens)
+                Common.ziggeo?.startAudioPlayer(tokens)
             } else {
                 Common.showAlertView("Audio recordings are empty.")
             }
@@ -169,7 +169,7 @@ class HomeViewController: UIViewController {
             }
             
             if (tokens.count > 0) {
-                Common.ziggeo?.showImages(tokens)
+                Common.ziggeo?.showImage(tokens)
             } else {
                 Common.showAlertView("Image recordings are empty.")
             }
@@ -296,10 +296,9 @@ extension HomeViewController: MenuActionDelegate {
     }
     
     func didSelectPlayVideoFromUrlMenu() {
-        Common.ziggeo?.play(fromUri: "http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/Sintel.mp4")
-//        Common.ziggeo?.play(fromUris: ["http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4",
-//                                       "http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/ForBiggerMeltdowns.mp4",
-//                                       "http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/Sintel.mp4",])
+        Common.ziggeo?.play(fromUri: ["http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4",
+                                      "http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/ForBiggerMeltdowns.mp4",
+                                      "http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/Sintel.mp4",])
     }
     
     func didSelectPlayLocalVideoMenu() {
@@ -317,7 +316,7 @@ extension HomeViewController: UIImagePickerControllerDelegate, UINavigationContr
     func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
         dismiss(animated: false) {
             let videoUrl = (info[.mediaURL] as! URL).path
-            Common.ziggeo?.play(fromUri: videoUrl)
+            Common.ziggeo?.play(fromUri: [videoUrl])
         }
     }
 }